Moved @Version to core project
