+++ router/src/main/java/io/micronaut/web/router/version/VersioningRouteMatchesFilter.java
@@ -17,6 +17,8 @@
 package io.micronaut.web.router.version;
 
 import io.micronaut.context.annotation.Requires;
+import io.micronaut.core.util.ArgumentUtils;
+import io.micronaut.core.version.annotation.Version;
 import io.micronaut.http.HttpRequest;
 import io.micronaut.web.router.UriRouteMatch;
 import io.micronaut.web.router.version.strategy.VersionExtractingStrategy;
